+++ pkg/alertmanager/state_replication.go
@@ -255,11 +255,11 @@ func (s *state) starting(ctx context.Context) error {
 
 // WaitReady is needed for the pipeline builder to know whenever we've settled and the state is up to date.
 func (s *state) WaitReady(ctx context.Context) error {
-	return s.Service.AwaitRunning(ctx)
+	return s.AwaitRunning(ctx)
 }
 
 func (s *state) Ready() bool {
-	return s.Service.State() == services.Running
+	return s.State() == services.Running
 }
 
 func (s *state) MergeGrafanaState(fs []*clusterpb.FullState) error {