+++ pkg/costattribution/manager_test.go
@@ -19,7 +19,7 @@ import (
 
 func newTestManager() *Manager {
 	logger := log.NewNopLogger()
-	limits, _ := testutils.NewMockCostAttributionLimits(0)
+	limits := testutils.NewMockCostAttributionLimits(0)
 	reg := prometheus.NewRegistry()
 	manager, err := NewManager(5*time.Second, 10*time.Second, logger, limits, reg)
 	if err != nil {
@@ -152,9 +152,7 @@ func TestManager_CreateDeleteTracker(t *testing.T) {
 	})
 
 	t.Run("Disabling user cost attribution", func(t *testing.T) {
-		var err error
-		manager.limits, err = testutils.NewMockCostAttributionLimits(1)
-		assert.NoError(t, err)
+		manager.limits = testutils.NewMockCostAttributionLimits(1)
 		assert.NoError(t, manager.purgeInactiveAttributionsUntil(time.Unix(11, 0)))
 		assert.Equal(t, 1, len(manager.sampleTrackersByUserID))
 
@@ -167,9 +165,7 @@ func TestManager_CreateDeleteTracker(t *testing.T) {
 	})
 
 	t.Run("Updating user cardinality and labels", func(t *testing.T) {
-		var err error
-		manager.limits, err = testutils.NewMockCostAttributionLimits(2)
-		assert.NoError(t, err)
+		manager.limits = testutils.NewMockCostAttributionLimits(2)
 		assert.NoError(t, manager.purgeInactiveAttributionsUntil(time.Unix(12, 0)))
 		assert.Equal(t, 1, len(manager.sampleTrackersByUserID))
 		assert.True(t, manager.SampleTracker("user3").hasSameLabels([]string{"feature", "team"}))
@@ -219,7 +215,7 @@ func TestManager_PurgeInactiveAttributionsUntil(t *testing.T) {
 
 	t.Run("Purge after inactive timeout", func(t *testing.T) {
 		// disable cost attribution for user1 to test purging
-		manager.limits, _ = testutils.NewMockCostAttributionLimits(1)
+		manager.limits = testutils.NewMockCostAttributionLimits(1)
 		assert.NoError(t, manager.purgeInactiveAttributionsUntil(time.Unix(5, 0)))
 
 		// User3's tracker should remain since it's active, user1's tracker should be removed