+++ pkg/ingester/activeseries/active_series_test.go
@@ -251,7 +251,7 @@ func (m *mockIndex) Series(ref storage.SeriesRef, builder *labels.ScratchBuilder
 }
 
 func TestActiveSeries_UpdateSeries_WithCostAttribution(t *testing.T) {
-	limits, _ := catestutils.NewMockCostAttributionLimits(0)
+	limits := catestutils.NewMockCostAttributionLimits(0)
 	reg := prometheus.NewRegistry()
 	manager, err := costattribution.NewManager(5*time.Second, 10*time.Second, log.NewNopLogger(), limits, reg)
 	require.NoError(t, err)